% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_channel.R
\name{add_channel}
\alias{add_channel}
\title{Add a channel to the gis_brick}
\usage{
add_channel(brck, channel)
}
\arguments{
\item{brck}{a RasterBrick as created by create_brick}

\item{channel}{a channel object as generated by create_channel}
}
\value{
a raster brick object as for brck but with additional layers populated
}
\description{
Reads the DEM file, sets up the raster brick and saves it
}
\details{
Populates the channel_id & channel_area layers. Alters the land_area layer to correspond.
}
