% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atb.R
\name{atb}
\alias{atb}
\title{Compute the topographic index}
\usage{
atb(project_path)
}
\arguments{
\item{project_path}{the path to the folder used for the analysis}
}
\description{
Compute the Topographic Index \code{log(area / tanb)}
}
\details{
The algorithm computes the topographic index from the upslope_area and tanb raster files in the project directory.
}
