% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borg.R
\name{borg}
\alias{borg}
\title{update a dynatopGIS class object}
\usage{
borg(old_object)
}
\arguments{
\item{old_object}{the current dynatopGIS object}

\item{is_class}{logical if the class based on this layer should be returned}
}
\value{
a RasterLayer of the requested information
}
\description{
Update a dynatopGIS object while keeping layers and channel
}
\details{
Setting is_class to TRUE and not providing a layer_name returns the overall classification
}
