% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn_in_class.R
\name{burn_in_class}
\alias{burn_in_class}
\title{Burn in distinct classes of hydrological response units (HRUs)}
\usage{
burn_in_class(brck, class_name, output_name, burns, json_path = getwd())
}
\arguments{
\item{brck}{a raster brick as generated by create_brick}

\item{class_name}{Name of the existing classification onto which to burn the new classes - must be a layer in brck}

\item{output_name}{name to give the output layer added to brck}

\item{burns}{list  Named list of layer in brck to burn into discretisation of HRUs that will be stamped onto the classification. Overrides any classification already defined}

\item{json_path}{Path of folder into which to write the json file identifying which burns were used.}
}
\value{
a raster layer as brck but with additional layer added.
}
\description{
Burn in distinct regions to a classification map
}
\details{
Create an additional layer in brck with a classification based on class_name with the burns overwriting the original class values. Burns are applied in the order theya are input.
}
