% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_atb.R
\name{calc_atb}
\alias{calc_atb}
\title{function to fill sinks}
\usage{
calc_atb(project_path, use_filled_dem = TRUE, max_iter = 10000)
}
\arguments{
\item{use_filled_dem}{logical, shouldthe filled DEM be used?}

\item{max_iterations}{Maximum number of iterations}
}
\description{
Compute the Topographic Index
}
\details{
The algorithm computes the area draining to a cell from those upstream presuming the flow is split between the downsteam (lower)  neighbouring cells proportional to the gradient.
}
