% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dem.R
\name{catchment_file}
\alias{catchment_file}
\alias{create_catchment}
\alias{check_catchment}
\alias{check_projection}
\title{Initialise the analysis by loading a dem}
\usage{
create_catchment(dem, fill_na = TRUE, ...)

check_catchment(ctch, req_names = req_catchment_layers())

check_projection(ctch, ...)
}
\arguments{
\item{dem}{must be either a rasterLayer or a file name}

\item{fill_na}{should NA values in dem be filled. See details}

\item{...}{passed to raster package functions to set options for reading files}

\item{catchment_file}{filename to save output to}

\item{stck}{a RasterStack or file name for testing to see if it is a catchment description}
}
\value{
A raster stack object with initialised layers or the name of file into which is is saved
}
\description{
Reads the DEM file, sets up the raster catchment description
}
\details{
Reads in the dem using the raster package. If fill_na all NA values other then those that link to the edge of the dem are filled with -Inf, so they can be identified as sinks.
}
