% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_properties.R
\name{compute_properties}
\alias{compute_properties}
\title{Function to compute statistics of brck pixels}
\usage{
compute_properties(brck, ...)
}
\arguments{
\item{brck}{rasterBrick or file containing the raster brick (as generated using create_brick)}

\item{...}{additional parameters passed to raster::brick if brck is a file name}
}
\description{
Computes statistics e.g. log(a/tanb) for raster cells
}
\details{
The algorithm works in two passes. The first computes the number of upstream pixels. The second sequences downslope to compute values.
}
