% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_channel.R
\name{create_channel}
\alias{create_channel}
\title{load a chanel file and ensure it has the correct properties}
\usage{
create_channel(
  shp,
  property_names = c(length = "length", startNode = "startNode", endNode = "endNode",
    width = "width"),
  river_file = "",
  default_width = 2,
  ...
)
}
\arguments{
\item{shp}{a spatialLinesDataFrame or path to file to be read by OGR to produce one}

\item{property_names}{named vector of columns of the spatial data frame to use for channel properties}

\item{river_file}{file to save the river network to if required}

\item{default_width}{default width of a channel if not specified in property_names. Defaults to 2 metres.}

\item{...}{passed to \code{raster::readOGR} if loading a file}
}
\value{
a channel SpatialLines object
}
\description{
Import channel data from an OGR file to the project directory
}
\details{
reading in the input file and covnerts it then saves a shape file of polygons with properties chanel_id,length,width,startNode,endNode in the project directory under the name \code{channel.shp}. Feild in the input file corresponding to the properties can be specified in the \code{property_names} vector.
}
