% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Create a dynamic TOPMODEL suitabel more model evaluation}
\usage{
create_model(stck, chn, hillslope_class, ...)
}
\arguments{
\item{stck}{a RasterBrick as created by create_brick}

\item{chn}{a channel object as created by create_channel}

\item{hillslope_class}{Name of the existing hillslope classification to use to generate the model - a layer in stck}

\item{...}{passed to \code{raster::stack} if loading a file}
}
\value{
Logical imdicating it has run. Outputs an rds file named after the classification in the project directory containing the model summary.
}
\description{
take a classification map and associatd GIS data then computes the GIS properties needed for dynamic TOPMODEL.
}
