% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downslope_pass.R
\name{downslope_pass}
\alias{downslope_pass}
\title{Function to compute upslope area}
\usage{
downslope_pass(brck, analysis_file = brck)
}
\arguments{
\item{dem}{File or raster containing dem data}
}
\description{
Computes a logical raster which identifies sinks
}
\details{
The algorithm computes sinks at all locations where it is the
lowest cell
}
