% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynatopGIS-package.R
\docType{package}
\name{dynatopGIS-package}
\alias{dynatopGIS-package}
\alias{_PACKAGE}
\title{dynatopGIS: Algorithms for Helping Build Dynamic TOPMODEL Implementations from Spatial Data}
\description{
A set of algorithms based on Quinn et al. (1991) \doi{10.1002/hyp.3360050106} for processing river network and digital elevation data to build implementations of Dynamic TOPMODEL, a semi-distributed hydrological model proposed in Beven and Freer (2001) \doi{10.1002/hyp.252}. The 'dynatop' package implements simulation code for Dynamic TOPMODEL based on the output of 'dynatopGIS'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://waternumbers.github.io/dynatopGIS/}
  \item \url{https://github.com/waternumbers/dynatopGIS}
  \item Report bugs at \url{https://github.com/waternumbers/dynatopGIS/issues}
}

}
\author{
\strong{Maintainer}: Paul Smith \email{paul@waternumbers.co.uk} (\href{https://orcid.org/0000-0002-0034-3412}{ORCID})

}
\keyword{internal}
