% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_min_ngh}
\alias{f_min_ngh}
\title{cpp wrapper function for passing up the catchments from river nodes}
\usage{
f_min_ngh(i, dem, offset)
}
\arguments{
\item{dem}{Digital elevation model as a vector}

\item{offset}{difference between index of neighbours and current cell - clockwise from top left}

\item{channel_id}{UID of channel in the pixel (id any) as a vector}
}
\value{
a list with the filled dem
}
\description{
cpp wrapper function for passing up the catchments from river nodes
}
