% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fN.R
\name{fN}
\alias{fN}
\title{Function to give neighbours of cells}
\usage{
fN(k, nr, nc, dx = NA)
}
\arguments{
\item{k}{cell index (only single values accepted)}

\item{nr}{number fo rows in original raster}

\item{nc}{number fo columnc in original raster}
}
\value{
the index of neighbouring cells
}
\description{
TO DO
}
\details{
This is in effect a cut down version of raster adjacent for use with the vectors exctraced by raster::getValues. See example below
}
\examples{
## create a test raster
rst <- raster(matrix(1:200,20,10))
## specify some locations
kvec <- c(1, # top left
          floor(nc/2), # top row middle column
          nc, # top right
          (nc*floor(nr/2))+1, # first column, middle row
          (nc*floor(nr/2))+floor(nc/2), # middle column middle row
          (nc*floor(nr/2))+nc, # last column, middle row
          (nc*(nr-1))+1, # bottom left
          (nc*nr)- floor(nc/2), # bottom row middle
          (nc*nr)) # bottom right

## test matching of locations
for(k in kvec){
    ra <- sort( raster::adjacent(rst,k,direction=8,pairs=FALSE) )
    fn <- sort( fN(k,nrow(rst),ncol(rst)) )
    if( length(ra)!=length(fn) || any( ra!=fn ) ){
        stop(paste("fN does not match raster adjacent for value",k))
    }
}
}
