% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_atb}
\alias{fun_atb}
\title{cpp wrapper function for computation of atb topographic wetness index}
\usage{
fun_atb(dem, area, is_channel, dist, max_iter)
}
\arguments{
\item{dem}{Digital elevation model}

\item{area}{Area of land surface in pixel}

\item{is_channel}{TRUE is a channel pixel}

\item{dist}{3x3 matrix of distances to each adjacent pixel}

\item{max_iter}{maximum number of iterations}
}
\value{
list of hillslope and channel properties
}
\description{
cpp wrapper function for computation of atb topographic wetness index
}
\details{
Computes topographic index area divided by tan(beta)
}
