% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_hru}
\alias{fun_hru}
\title{cpp wrapper function for computation of redistribution matrices}
\usage{
fun_hru(dem, grad, land_area, channel_area, channel_id, hillslope_id,
  offset, dx, cl, max_index)
}
\arguments{
\item{dem}{Digital elevation model}

\item{land_area}{Area of land surface in pixel}

\item{hillslope}{hillslope class of each pixel}

\item{channel}{channel class of each pixel}

\item{number_hillslope_class}{number of hill slope classes}

\item{number_channel_class}{number of channel classes}

\item{dist}{3x3 matrix of distances to ajoining cells}
}
\value{
list of hillslope and channel properties
}
\description{
cpp wrapper function for computation of redistribution matrices
}
