% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_increase_order}
\alias{fun_increase_order}
\title{cpp wrapper function for filling of sinks}
\usage{
fun_increase_order(dem, order, co, nc)
}
\arguments{
\item{dem}{Digital elevation model}

\item{is_channel}{TRUE is a channel pixel}

\item{delta}{3x3 matrix of minimum evelation drop to each adjacent pixel}

\item{max_iter}{maximum number of iterations}
}
\value{
matrix containing filled dem
}
\description{
cpp wrapper function for filling of sinks
}
