% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_is_sink}
\alias{fun_is_sink}
\title{cpp wrapper function for filling of sinks}
\usage{
fun_is_sink(dem, nc)

fun_is_sink(dem, nc)
}
\arguments{
\item{dem}{Digital elevation model}

\item{is_channel}{TRUE is a channel pixel}

\item{delta}{3x3 matrix of minimum evelation drop to each adjacent pixel}

\item{max_iter}{maximum number of iterations}

\item{dem}{Digital elevation model}

\item{is_channel}{TRUE is a channel pixel}

\item{delta}{3x3 matrix of minimum evelation drop to each adjacent pixel}

\item{max_iter}{maximum number of iterations}
}
\value{
matrix containing filled dem

matrix containing filled dem
}
\description{
cpp wrapper function for filling of sinks

cpp wrapper function for filling of sinks
}
