% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_redistribution}
\alias{fun_redistribution}
\title{cpp wrapper function for computation of redistribution matrices}
\usage{
fun_redistribution(dem, land_area, hillslope, channel,
  number_hillslope_class, number_channel_class, dist)
}
\arguments{
\item{dem}{Digital elevation model}

\item{land_area}{Area of land surface in pixel}

\item{hillslope}{hillslope class of each pixel}

\item{channel}{channel class of each pixel}

\item{number_hillslope_class}{number of hill slope classes}

\item{number_channel_class}{number of channel classes}
}
\value{
list of hillslope and channel properties
}
\description{
cpp wrapper function for computation of redistribution matrices
}
