% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_sink_fill}
\alias{fun_sink_fill}
\title{cpp wrapper function for filling of sinks}
\usage{
fun_sink_fill(dem, is_channel, delta, max_iter)
}
\arguments{
\item{dem}{Digital elevation model}

\item{is_channel}{TRUE is a channel pixel}

\item{delta}{3x3 matrix of minimum evelation drop to each adjacent pixel}

\item{max_iter}{maximum number of iterations}
}
\value{
matrix containing filled dem
}
\description{
cpp wrapper function for filling of sinks
}
