% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_tanb}
\alias{fun_tanb}
\title{cpp wrapper function for computing average gradient}
\usage{
fun_tanb(dem, is_channel, dist)
}
\arguments{
\item{dem}{Digital elevation model}

\item{is_channel}{TRUE if pixel contains a channel}

\item{dist}{3x3 matrix of distance to adjacent pixels}
}
\value{
matrix of average gradients
}
\description{
cpp wrapper function for computing average gradient
}
