% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_upslope_area}
\alias{fun_upslope_area}
\title{cpp wrapper function for computation of upslope area}
\usage{
fun_upslope_area(dem, area, is_channel, dist, max_iter)
}
\arguments{
\item{dem}{Digital elevation model}

\item{area}{Area of land surface in pixel}

\item{is_channel}{TRUE if pixel contains a channel}

\item{dist}{3x3 matrix of distance to adjacent pixels}

\item{max_iter}{maximum number of iterations}
}
\value{
matrix of upslope area
}
\description{
cpp wrapper function for computation of upslope area
}
