% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun_upslope_pass}
\alias{fun_upslope_pass}
\title{cpp wrapper function for passing up the catchments from river nodes}
\usage{
fun_upslope_pass(dem, order, offset)
}
\arguments{
\item{dem}{Digital elevation model as a vector}

\item{order}{initial order values as a vector, internally starts at one and move upstream}

\item{nc}{number of columns in the matrix}
}
\value{
a list with the filled dem and order
}
\description{
cpp wrapper function for passing up the catchments from river nodes
}
