% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dem.R
\name{gis_brick}
\alias{gis_brick}
\alias{create_brick}
\alias{check_brick}
\title{Initialise the analysis by loading a dem}
\usage{
create_brick(dem, fill_na = TRUE, ...)

check_brick(brck)
}
\arguments{
\item{dem}{must be either a rasterLayer or a file name}

\item{fill_na}{should NA values in dem be filled. See details}

\item{...}{passed to raster::raster}

\item{brck}{a RasterBrick for testing to match those created by create_brick}
}
\value{
raster brick object with initialised layers
}
\description{
Reads the DEM file, sets up the raster brick
}
\details{
Reads in the dem using the raster package. Fi fill_na all NA values other then those that link to the edge of the dem are filled with -Inf, so they can be identified as sinks.
}
