% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{glist_to_raster}
\alias{glist_to_raster}
\title{Function to convert to and from a catchment object}
\usage{
glist_to_raster(glist, layer = "dem")
}
\description{
Function to convert to and from a catchment object
}
