% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise}
\alias{initialise}
\title{Initialise the analysis by loading a dem}
\usage{
initialise(dem, channel, analysis_file = "", fill_na = TRUE, ...)
}
\arguments{
\item{analysis_file}{file to save the analysis to}

\item{dem_file}{name of file to load as the dem}
}
\value{
raster brick object pointing to the analysis file
}
\description{
Reads the DEM file, sets up the raster brick and saves it
}
\details{
Currently reads and then writes file using the raster package. No checks as to the projection are enforced. Output is project_path/dem.tif
}
