% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sink.R
\name{is_sink}
\alias{is_sink}
\title{Function to compute upslope area}
\usage{
is_sink(dem)
}
\arguments{
\item{dem}{File or raster containing dem data}
}
\description{
Computes a logical raster which identifies sinks
}
\details{
The algorithm computes sinks at all locations where it is the
lowest cell
}
