% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_channel.R
\name{load_channel}
\alias{load_channel}
\title{load a chanel file and ensure it has the correct properties}
\usage{
load_channel(file_name, property_names, project_path = ".",
  default_width = 2)
}
\arguments{
\item{file_name}{name of file to read in}

\item{property_names}{named vector of columns of the spatial data frame to use for channel properties}
}
\value{
A SpatialLinesDataFrame with properties chanel_id,length,width,startNode,endNode
}
\description{

}
