% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dem.R
\name{load_dem}
\alias{load_dem}
\title{Load a dem and compute its area}
\usage{
load_dem(dem_file, project_path = ".")
}
\arguments{
\item{dem_file}{name of file to load as the dem}

\item{project_path}{the path of the project to which to write dem file}
}
\value{
TRUE if successful
}
\description{
Reads the DEM file and writes it to the project directory
}
\details{
Currently reads and then writes file using the raster package. No checks as to the projection are enforced. Output is project_path/dem.tif
}
