% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hru}
\alias{rcpp_hru}
\title{cpp wrapper function for computation of hru properties}
\usage{
rcpp_hru(
  dem,
  hillslope_id,
  channel_id,
  land_area,
  channel_area,
  grad,
  atb,
  W,
  area,
  s_bar,
  atb_bar,
  rst_prop
)
}
\arguments{
\item{dem}{Digital elevation model}

\item{hillslope_id}{hillslope class of each pixel}

\item{channel_id}{UID of channel present in the pixel}

\item{land_area}{Area of land surface in pixel}

\item{channel_area}{Area of channel surface in pixel}

\item{grad}{gradient of the cell}

\item{atb}{topographic index ln(upslope area / tan beta) of the cell}

\item{W}{saturated zone redistribution matrix}

\item{area}{HSU area}

\item{s_bar}{HSU average gradient}

\item{atb_bar}{HSU averable topographic index}

\item{rst_prop}{dimension ans resolution of raster converted to vectors}
}
\description{
cpp wrapper function for computation of hru properties
}
