% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_sink_fill}
\alias{rcpp_sink_fill}
\title{cpp wrapper function for passing up the catchments from river nodes}
\usage{
rcpp_sink_fill(dem, channel_id, offset)
}
\arguments{
\item{dem}{Digital elevation model as a vector}

\item{channel_id}{UID of channel in the pixel (id any) as a vector}

\item{offset}{difference between index of neighbours and current cell - clockwise from top left}
}
\value{
a list with the filled dem
}
\description{
cpp wrapper function for passing up the catchments from river nodes
}
