% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dem.R
\name{req_catchment_layers}
\alias{req_catchment_layers}
\title{returns the list of required layers in the catchment file}
\usage{
req_catchment_layers()
}
\description{
returns the list of required layers in the catchment file
}
