% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_fill.R
\name{sink_fill}
\alias{sink_fill}
\title{function to fill sinks}
\usage{
sink_fill(project_path, max_iter = 10, minimum_tangent = 0.001,
  use_filled_dem = FALSE)
}
\arguments{
\item{project_path}{folder whcih is being used for the analysis}

\item{max_iter}{maximum number of iterations of the algorithm}

\item{minimum_tangent}{minimum value of teh gradient to the next cell downstream}

\item{use_filled_dem}{logical indicating if the analysis should start with the filled DEM}
}
\description{
Fills sinks so that they have a slope of at least the minimum_tangent to the downstream cell
}
\details{
the algotithm works by looping over all the cells in the DEM and is relavitley robust but inefficent. If the maximum number of iterations is exceeded and there are still sinks the \code{use_filled_dem} flag can be used to restart from the final state.
}
