% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{split_to_class}
\alias{split_to_class}
\title{Discrete a catchment into hydrological response units (HRUs)}
\usage{
split_to_class(project_path, output_name, cuts)
}
\arguments{
\item{output_name}{name to give the output files}

\item{cuts}{A named list of cuts of make to form the HRU. Names should correspond to raster layers in the project directory. Values should be numeric and define either the number of bands (single value) or breaks between band (multiple values)}

\item{project_dir}{Directory of the project}

\item{class_name}{Name fo the existing classification onto which to burn the new classes}

\item{burns}{list  Named list of geometries (supplied as rasters) to burn into discretisation of HRUs that will be stamped onto the classification. Overrides any classification already defined}
}
\value{
Logical imdicating it has run. Outputs a raster of classifications and a json of HRU properties to the project directory.
}
\description{
Discretise a catchment into a set hydrological response units (HRUs) according to any number of landscape layers and cuts
}
\details{
This applies the given cuts to the supplied landscape layers to produce areal groupings of the catchment. These are numbered using a cantor pairing scheme. In theory you could reverse out the class of each layer if required but this isn't implimented.
}
