% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanb.R
\name{tanb}
\alias{tanb}
\title{Function to compute tanb}
\usage{
tanb(project_path, use_filled_dem = TRUE)
}
\arguments{
\item{project_path}{folder which is being used for the analysis}

\item{use_filled_dem}{logical, shouldthe filled DEM be used?}
}
\description{
Computes the average downslope gradient for use in topographic index calculations
}
\details{
The algorithm computes the area draining to a cell from those upstream presuming the flow is split between the downsteam (lower)  neighbouring cells proportional to the gradient.
}
