% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upslope_area.R
\name{upslope_area}
\alias{upslope_area}
\title{Function to compute upslope area}
\usage{
upslope_area(project_path, use_filled_dem = TRUE, max_iter = 10000)
}
\arguments{
\item{project_path}{folder whcih is being used for the analysis}

\item{use_filled_dem}{logical, shouldthe filled DEM be used?}

\item{max_iter}{Maximum number of iterations}
}
\description{
Computes the upslope area of a pixel for use in topographic index calculations
}
\details{
The algorithm computes the area draining to a cell from those upstream presuming the flow is split between the downsteam (lower)  neighbouring cells proportional to the gradient.
}
