% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_channel.R
\name{add_channel}
\alias{add_channel}
\title{Add a channel to the DEM}
\usage{
add_channel(project_path)
}
\arguments{
\item{project_path}{the path to the folder used for the analysis}
}
\description{
Adds a chanel to the DEM by working out the land and channel areas and channel id
}
\details{
intersects the DEM and river network to determine which of the raster DEM cells contain parts of the river network. From this three raster maps are produced containing:
* the area in the DEM cell covered by land
* the area in the DEM cell covered by channel
* the id of the channel
If multiple river lengths intersect a DEM cell the id of that with the largest intersection is used.
}
