% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borg.R
\name{borg}
\alias{borg}
\title{update a dynatopGIS class object}
\usage{
borg(old_object, force = FALSE)
}
\arguments{
\item{old_object}{the current dynatopGIS object}

\item{force}{force conversion even to an older version}
}
\value{
an object of class dynatopGIS updated to the current version of the methods
}
\description{
Update a dynatopGIS object while keeping layers, channel and classification
}
\details{
Currently the model is not copied
}
