% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn_in_class.R
\name{burn_in_class}
\alias{burn_in_class}
\title{Burn in distinct classes of hydrological response units (HRUs)}
\usage{
burn_in_class(brck, class_name, output_name, burns)
}
\arguments{
\item{class_name}{Name of the existing classification onto which to burn the new classes}

\item{output_name}{name to give the output files}

\item{burns}{list  Named list of geometries (supplied as rasters) to burn into discretisation of HRUs that will be stamped onto the classification. Overrides any classification already defined}

\item{project_path}{folder which is being used for the analysis}
}
\value{
Logical indicating it has run. Outputs a raster of classifications and a json of HRU properties to the project directory.
}
\description{
Burn in distinct regions to a classification map
}
\details{
Replaces exisiting classes with those from the specified classification maps to be burnt in. burns are applied in the order theya are input.
}
