% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Create a dynamic TOPMODEL suitabel more model evaluation}
\usage{
create_model(project_path, hillslope_class)
}
\arguments{
\item{hillslope_class}{Name of the existing hillslope classification to use to generate the model}

\item{project_dir}{directory of the project being worked on}
}
\value{
Logical imdicating it has run. Outputs an rds file named after the classification in the project directory containing the model summary.
}
\description{
take a classification map and associatd GIS data then computes the GIS properties needed for dynamic TOPMODEL
}
