% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_channel.R
\name{load_channel}
\alias{load_channel}
\title{load a chanel file and ensure it has the correct properties}
\usage{
load_channel(file_name, property_names, project_path = ".",
  default_width = 2)
}
\arguments{
\item{file_name}{name of file to read in}

\item{property_names}{named vector of columns of the spatial data frame to use for channel properties}

\item{project_path}{folder which is being used for the analysis}

\item{default_width}{default width of a channel if not specified in property_names. Defaults to 2 metres.}
}
\value{
TRUE if runs otherwise an error message.
}
\description{
Import channel data from an OGR file to the project directory
}
\details{
reading in the input file and covnerts it then saves a shape file of polygons with properties chanel_id,length,width,startNode,endNode in the project directory under the name \code{channel.shp}. Feild in the input file corresponding to the properties can be specified in the \code{property_names} vector.
}
