% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hru}
\alias{rcpp_hru}
\title{cpp wrapper function for computation of redistribution matrices}
\usage{
rcpp_hru(
  dem,
  grad,
  land_area,
  channel_area,
  channel_id,
  hillslope_id,
  offset,
  dx,
  cl,
  max_index
)
}
\arguments{
\item{dem}{Digital elevation model}

\item{grad}{average downslope gradient}

\item{land_area}{Area of land surface in pixel}

\item{channel_area}{Area of channel surface in pixel}

\item{channel_id}{UID of channel present in the pixel}

\item{hillslope_id}{hillslope class of each pixel}

\item{offset}{- difference between cell index of adjacent cells and current cell index - clockwise from top left}

\item{dx}{distance between cell centres - from top left in clockwise direction}

\item{cl}{contour length - from top left in a clockwise direction. The 9th value is used for cells split beteen land and channel}

\item{max_index}{maximum value of the hillslope and channel id's}
}
\value{
list of hillslope and channel properties
}
\description{
cpp wrapper function for computation of redistribution matrices
}
