% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_fill.R
\name{sink_fill}
\alias{sink_fill}
\title{Fill the sinks with a dem}
\usage{
sink_fill(stck, ...)
}
\arguments{
\item{stck}{a rasterBrick object or whatever}

\item{...}{additional varaibles passed to raster::brick if stck is a file name}
}
\value{
a rasterBrick object
}
\description{
TO DO
}
\details{
TO DO
}
