% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_fill.R
\name{sink_fill}
\alias{sink_fill}
\title{Fill the sinks with a dem}
\usage{
sink_fill(stck, max_it = 1e+06, ...)
}
\arguments{
\item{stck}{a RasterStack object as generated by create_catchment}

\item{max_it}{maximum number of replacements to trial}

\item{...}{additional variables passed to raster::stack if stck is a file name}
}
\value{
an altered RasterStack object with the filled_dem populated.
}
\description{
TO DO
}
\details{
Sinks are then identified as either having an Inf dem values or by not having a lower neighbour. Sinks are replaced in order of the lowest neighbouring value (min first) by the mean of the adjacent finite value cells. See Vignette for a more detailed desciption.

If stck is a character object it is presemed to be the path to a file containing the catchment RasterStack. If so it read in and overwritten with the result using the raster package.
}
