% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_fill.R
\name{sink_fill}
\alias{sink_fill}
\title{Fill the sinks with a dem}
\usage{
sink_fill(ctch, min_grad = 1e-04, max_it = 1e+06, verbose = FALSE, ...)
}
\arguments{
\item{min_grad}{Minimum gradient between cell centres}

\item{max_it}{maximum number of replacements to trial}

\item{verbose}{pront out additional diagnostic information}

\item{...}{additional variables passed to raster::stack if stck is a file name}

\item{stck}{a RasterStack object as generated by create_catchment}
}
\value{
an altered RasterStack object with the filled_dem populated.
}
\description{
The sink filling algorithm of Planchona and Darboux 2001
}
\details{
The algorithm implimented in Planchona and Darboux, "A fast, simple and versatile algorithm to fill the depressions in digital elevation models" Catena 46 (2001). A pdf can be found at https://horizon.documentation.ird.fr/exl-doc/pleins_textes/pleins_textes_7/sous_copyright/010031925.pdf.

If stck is a character object it is presemed to be the path to a file containing the catchment RasterStack. If so it read in and overwritten with the result using the raster package.
}
