% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_to_class.R
\name{split_to_class}
\alias{split_to_class}
\title{Split a catchment into hydrological response units (HRUs)}
\usage{
split_to_class(stck, split_name, cuts, json_path = getwd())
}
\arguments{
\item{stck}{a RasterStack as created by create_catchment}

\item{split_name}{name to give the split}

\item{cuts}{A named list of cuts of make to form the HRU. Names should correspond to raster layers in the project directory. Values should be numeric and define either the number of bands (single value) or breaks between band (multiple values)}

\item{json_path}{the folder to write out json file containing the summary of splits to}
}
\value{
as for stck but containing an additional layer with the split classes
}
\description{
Split a catchment into a set hydrological response units (HRUs) according to any number of landscape layers and cuts
}
\details{
This applies the given cuts to the supplied landscape layers to produce areal groupings of the catchment. These are numbered using a cantor pairing scheme. In theory you could reverse out the class of each layer if required but this isn't implimented.
}
